requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tenacity>=8.2.3  # For retry mechanisms
aiohttp>=3.9.1   # For potential future async implementations